--- a/cosmos/__init__.py
+++ b/cosmos/__init__.py
@@ -5,7 +5,7 @@ Astronomer Cosmos is a library for rendering dbt workflows in Airflow.
 
 Contains dags, task groups, and operators.
 """
-__version__ = "1.5.0a1"
+__version__ = "1.5.0a2"
 
 
 from cosmos.airflow.dag import DbtDag