Introduce DAG level cache. Based on initial tests, this can improve the performance of running a Cosmos-powered DAG by 59% (validated with the repo's `basic_cosmos_dag`).

To enable this feature, users should set the environment variable `AIRFLOW__COSMOS__EXPERIMENTAL_CACHE=1`.

What is missing (why this is draft PR):
* Allow users to customise their own `calculate_current_version`
* Refactor so there isn't so much code duplicate between `DbtDag` and `TaskGroup`
* Allow users to purge the cache (via env var or param passed in the Airflow UI - second option is probably preferable)
* Unify `_create_cache_identifier` and `create_cache_identifier_v2`
* Write tests
* The cache should be invalidated if there is an error unpickling

This will be done in a separate PR:
* have a way of storing cache remotely (to be addressed in a separate ticket: https://github.com/astronomer/astronomer-cosmos/issues/927). There is a concern about how secure is pickle, so we may need to cache representation.
* button in the UI to purge cache for individual DAG/TaskGroup
* button in the UI to purge all Cosmos cache
* enable feature by default before 1.5 release (TODO: log ticket)

Closes: #926
Closes: #990

**How to reproduce**

Run using the new cache feature:
```
time AIRFLOW__COSMOS__ENABLE_CACHE=1 AIRFLOW__COSMOS__EXPERIMENTAL_CACHE=1 airflow dags test basic_cosmos_dag  `date -Iseconds`
```
Ran in `7.03s user 2.59s system 81% cpu 11.808 total`
<img width="1506" alt="Screenshot 2024-05-21 at 17 28 27" src="https://github.com/astronomer/astronomer-cosmos/assets/272048/4b2974e9-3cf9-4161-b442-229108e1fc6b">

Run without the new cache feature:
```
time  AIRFLOW__COSMOS__ENABLE_CACHE=1 AIRFLOW__COSMOS__EXPERIMENTAL_CACHE=0 airflow dags test basic_cosmos_dag  `date -Iseconds`
```
Ran in `18.75s user 2.62s system 73% cpu 28.898 total`
<img width="1504" alt="Screenshot 2024-05-21 at 17 28 39" src="https://github.com/astronomer/astronomer-cosmos/assets/272048/cbf9c81e-33ca-472c-8bf0-0bb62ce52a3b">