@tatiana 2024-06-03
When using the current approach in a distributed environment, there are two challenges:

1. Concurrent tasks (in the same node) try to create the cache at the same time
2. Tasks running in different nodes (that didn't have the cache) have to generate it

We'll look into improving this.

Examples of the behavior in a distributed Airflow environment:

![Image](https://github.com/astronomer/astronomer-cosmos/assets/272048/884832ce-e1d1-49f1-97b0-b4888f212bce)
![Image](https://github.com/astronomer/astronomer-cosmos/assets/272048/f384a182-eefd-4869-9abd-0433aeea2ea7)