@tatiana 2024-06-05
For this approach to work in a distributed Airflow environment, the cache would have to be stored remotely.

storing a pickle in a remote location (e.g., an object store) has security implications. For example, someone could provide malicious shell code as input, causing remote code execution. For this reason, I'm closing this PR.

I'm closing this PR in favor of a more promising dbt ls caching strategy in #1014.